further code cleanup
