+++ src/couchdb/dtos.ts
