+++ src/report-changes/report-changes.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { CouchdbReportChangesService } from './core/couchdb-report-changes.service';
+import { ReportChangesService } from './core/report-changes.service';
 import { CouchdbChangesService } from './storage/couchdb-changes.service';
 import { NotificationModule } from '../notification/notification.module';
 import { ReportModule } from '../report/report.module';
@@ -11,10 +11,10 @@ import { TestController } from './test-controller';
   controllers: [TestController],
   imports: [NotificationModule, ReportModule, HttpModule],
   providers: [
-    CouchdbReportChangesService,
+    ReportChangesService,
     CouchdbChangesService,
     CouchDbClient, // TODO: pack this into a CouchDbModule together with HttpModule import etc.
   ],
-  exports: [CouchdbReportChangesService],
+  exports: [ReportChangesService],
 })
 export class ReportChangesModule {}
