+++ src/report-changes/storage/couchdb-changes.service.ts
@@ -11,6 +11,7 @@ import {
   expand,
   filter,
   finalize,
+  map,
   mergeMap,
   Observable,
   of,
@@ -21,13 +22,20 @@ import {
 } from 'rxjs';
 import { CouchDbClient } from '../../couchdb/couch-db-client.service';
 import { CouchDbChangesResponse } from '../../couchdb/dtos';
+import {
+  DatabaseChangeResult,
+  DatabaseChangesService,
+} from './database-changes.service';
 
+/**
+ * Access _changes from a CouchDB
+ */
 @Injectable()
-export class CouchdbChangesService {
+export class CouchdbChangesService extends DatabaseChangesService {
   // TODO: centralize this config by refactoring couchdbClient and providing configured clients through DI
   // TODO: check if this is the correct db for our changes from app
   private dbUrl: string = this.configService.getOrThrow('DATABASE_URL');
-  private databaseName: string = 'app'; // TODO: move to config and clean up .env, clarifying different DBs there
+  private databaseName = 'app'; // TODO: move to config and clean up .env, clarifying different DBs there
   private databaseUser: string = this.configService.getOrThrow('DATABASE_USER');
   private databasePassword: string =
     this.configService.getOrThrow('DATABASE_PASSWORD');
@@ -38,20 +46,21 @@ export class CouchdbChangesService {
     private couchdbClient: CouchDbClient,
     private configService: ConfigService,
   ) {
+    super();
     const authHeader = Buffer.from(
       `${this.databaseUser}:${this.databasePassword}`,
     ).toString('base64');
     this.authHeaderValue = `Basic ${authHeader}`;
   }
 
-  private changesSubj = new ReplaySubject<CouchDbChangesResponse>(1);
+  private changesSubj = new ReplaySubject<DatabaseChangeResult[]>(1);
   private changesSubscription: Subscription | undefined;
 
-  subscribeToAllNewChanges(): Observable<CouchDbChangesResponse> {
+  subscribeToAllNewChanges(): Observable<DatabaseChangeResult[]> {
     if (!this.changesSubscription) {
       let lastSeq = 'now';
       const changesFeed = of({}).pipe(
-        mergeMap((_) => this.fetchChanges(lastSeq, true)),
+        mergeMap(() => this.fetchChanges(lastSeq, true)),
         filter((res) => res.last_seq !== lastSeq),
         tap((res) => (lastSeq = res.last_seq)),
         // poll regularly to get latest changes
@@ -59,7 +68,9 @@ export class CouchdbChangesService {
         tap((res) => console.log(res)),
       );
 
-      this.changesSubscription = changesFeed.subscribe(this.changesSubj);
+      this.changesSubscription = changesFeed
+        .pipe(map((res) => res.results))
+        .subscribe(this.changesSubj);
     }
 
     return this.changesSubj.asObservable().pipe(
@@ -80,8 +91,8 @@ export class CouchdbChangesService {
    * @param getAllPending Whether to trigger multiple requests and emit multiple values before completing, in case the first request has more pending changes
    */
   fetchChanges(
-    since: string = 'now',
-    getAllPending: boolean = false,
+    since = 'now',
+    getAllPending = false,
   ): Observable<CouchDbChangesResponse> {
     return this.couchdbClient
       .changes(this.dbUrl, this.databaseName, {
