+++ src/report-changes/storage/database-changes.service.ts
@@ -1,0 +1,12 @@
+import { Observable } from 'rxjs';
+import { CouchDbChangeResult } from '../../couchdb/dtos';
+
+/**
+ * Provides access to a stream of document changes for a database.
+ */
+export abstract class DatabaseChangesService {
+  abstract subscribeToAllNewChanges(): Observable<DatabaseChangeResult[]>;
+}
+
+// TODO: don't expose the CouchDb specific changes interface, map this to domain interfaces (maybe use DocChangeDetails)?
+export type DatabaseChangeResult = CouchDbChangeResult;
